creating the api for sign up and creating the front end of signup for the user to login. and Implementing snack bar or notifications
creating the signup api route for the user to be able to register themselves in the mainRouter.js file of the Router.js folder.

the route is router.post('/signup', signUp);
the controller code is -> const userCredential = require("../models/userSchema.js");
Modified the UI
Added logo
implemented snackbar notifications for the login status update in the app.

Implemented a front-end "Login to have a personalized experience." text on the login page with a typewriter effect.